+++ cfdm/docstring/docstring.py
@@ -1325,13 +1325,13 @@ _docstring_substitution_definitions = {
 
                 If False then do not show such data elements, *unless
                 data elements have been previously cached*, thereby
-                avoiding a potential computational cost.
+                avoiding a potentially high computational cost.
 
                 If `None` (the default) then the value of *data* will
                 taken from the `{{package}}.display_data` function.
 
                 Note that whenever data elements are displayed, they
-                will be cached for fast future retrieval.""",
+                are cached for fast future retrieval.""",
     # ----------------------------------------------------------------
     # Method description substitutions (4 levels of indentation)
     # ----------------------------------------------------------------